4x4 Keypad with LCD
This code is for a DIY calculator using an LCD, a keypad, and an Arduino. It performs basic arithmetic operations like addition, subtraction, multiplication, and division. The code also handles button detection, result calculation, and displaying results on the LCD.

Check Address of your LCD and your Connections before uploading the code.